+++ client/src/components/Calendar/index.jsx
@@ -1,32 +1,33 @@
-import React, { useEffect } from 'react';
+import React, { useEffect, useState } from 'react';
 import FullCalendar from '@fullcalendar/react';
 import dayGridPlugin from '@fullcalendar/daygrid';
 import interactionPlugin from '@fullcalendar/interaction';
 import { useDispatch, useSelector } from 'react-redux';
 import styled from 'styled-components';
 import Modal from '../Modal';
-import { setTaskOpen, getTask } from '../../state';
+import { setTaskOpen } from '../../state';
 import axios from 'axios';
 
 export default function Calender() {
-  const { tasks } = useSelector((state) => state.tasks.value);
   const { openTask } = useSelector((state) => state.modal.value);
   const dispatch = useDispatch();
 
-  // useEffect(() => {
-  //   const handleData = async () => {
-  //     try {
-  //       const Response = await axios.get(`/api/v1/allTasks`);
-  //       if (Response.status === 200) {
-  //         dispatch(getTask(Response.data.data));
-  //         setCount(Response.data.data[0].totaltask);
-  //       }
-  //     } catch (error) {
-  //       console.log(error, 'error');
-  //     }
-  //   };
-  //   handleData();
-  // }, [dispatch, pagination]);
+  const [tasks, setTask] = useState([]);
+
+  useEffect(() => {
+    const handleData = async () => {
+      try {
+        const Response = await axios.get(`/api/v1/allTasks`);
+        console.log(Response);
+        if (Response.status === 200) {
+          setTask(Response.data.data);
+        }
+      } catch (error) {
+        console.log(error, 'error');
+      }
+    };
+    handleData();
+  }, [dispatch]);
 
   const dataa = () => {
     const newd = tasks.map((task) => {
