+++ server/src/database/quieres/tasks/index.ts
@@ -1,11 +1,10 @@
 import addTaskQuery from './addTaskQuery';
-import getTasksQuery from './getTasksQuery';
 import getTasksSectionQuery from './getTasksSectionQuery';
 import editTaskQuery from './editTaskQuery';
 import deleteTaskQuery from './deleteTaskQuery';
 import addUserTaskQuery from './addUserTaskQuery';
 import dragTaskQuery from './dragTaskQuery';
-import getAllTasksQuery from './getAllTasksQuery';
+import { getAllTasksQuery, getTasksQuery } from './getTasksQuery';
 
 export {
   addTaskQuery,
